Changes the constructor by adding a reference to the server object and adds catalog information in the show command
The backup_id parameter has been removed from the constructor (never used). A new mandatory parameter has been added and it is a reference to the server object (to which the backup belongs). This allows to perform queries in the server catalog like previous or next base backup. This is the first step in order to allow WAL file size calculations.