Adds catalog information
Functions that help retrieving catalog information have been added, such as 'get_available_backups' which returns a dictionary of valid backup objects (backup ID is the key). Added functions that, given a backup ID, get the previous and next backup in the server catalog. Added the new constructor of the the Backup class.