Retention status and report() method for retention policies
Created the following states for retention policies:
VALID, OBSOLETE, NONE and POTENTIALLY_OBSOLETE (an
object which is OBSOLETE but cannot be removed
automatically due to minimum_redundancy requirements).
Created the report() method for the retention policy
base class, which exected the _backup_report() method
for base backups and the _wal_report() method for WAL
retention policies (currently not enforced).
The report method iterates through the DONE backups
and according to the retention policy, classifies
the backup. RedundancyRetentionPolicy uses the number
of backups, RecoveryWindowRetentionPolicy uses the
time window and the recoverability point concept.