27: Addresses potential corruption of WAL xlog.db files.
In barman.lockfile.release() the file is unlinked (deleted). This effectively
nullifies any future attempts to lock the file by a blocking process by deleting
the open file table entry upon which the flock is based.

This commit removes the unlink and instead unlocks the file and then closes the file
descriptor leaving the lock file and open file table entry intact.