Store deduplication effects for incremental backup
When incremental backup is enabled and uses hard links
(reuse_backup = link), output of 'backup' command reports
the effects of deduplication. The metrict is stored along
the backup.info file in the 'deduplicated_size' field.

IMPORTANT: this metric refers to the increment in size
of the current backup from the previous backup and reflects
only the situation at backup time.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>