Improved cache management of backups
Streamlined the management of the cache of the backups.
It is now possible to register and remove a backup from the cache.
The cache structure has been simplified and now it is a simple
dictionary of backups. Status filters are applied by the
get_available_backups() method.
Managed registration and removal of a backup in the cache
during backup operations.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>