Forbid the delete of a running backup
Block the execution of a barman delete command on a backup that is in progress.
Use locks to ensure that there are no running backups on that server.
Otherwise check for the position of the backup to be deleted in the
catalogue and its status: if it is the first backup and its status is STARTED or EMPTY,
the backup is running and delete is forbidden.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>