Pin (temporarily) pytest-catchlog to version 1.1
The pytest-catchlog release 1.2.0 broke Python 2.6 compatibility, so pin
the 1.1 version until it will be restored.

Ref: https://github.com/eisensheng/pytest-catchlog/pull/15

In latest release of pytest-catchlog the result of caplog.records() call
is a copy of the event list so any modification of that has no global
effect. Change the code that uses caplog.records() to not rely on the
previous undocumented behaviour.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>