Add support for 'streaming_archiver' option
Add a global/server boolean option called 'streaming_archiver' which
enables the use of the PostgreSQL's streaming protocol to receive
transaction logs from a server.
If set to 'on', Barman expects to find `pg_receivexlog` in the PATH
(see `path` option) and that streaming connection for the server is
properly working. Subsequently, it activates both proper connection
checks and management of WAL files.

Refactored the code to use a new class, the StreamingWalArchiver,
which is responsible for managing WAL files using a streaming
connection.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>