Add support for 'archiver' option
Add a global/server boolean option called 'archiver' which
enables continuous WAL archiving (through log shipping using
PostgreSQL's archive_command) for a specific server.
If set the 'true' (default), Barman expects that PostgreSQL
continously deposits WAL files in the 'incoming' directory
of that server, and subsequently activates proper checks and
management.

Refactored the code to use a new class, the FileWalArchiver.
The class is in charge of both archival and compression of WAL files.
This is the first step towards the implementation of streaming
archiving support.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>