remove redundant sete[gu]id() calls
Python's documentation of the os.setuid/os.setgid calls is
oversimplified to the point of being wrong. Under the hood,
the setuid(2)/setgid(2) functions are used, thus making these
calls as POSIX compliant as the underlying operating system.
As barman won't be installed SUID (it's an interpreter file),
effective, real and saved user id should always be the same.
That way, the only reason to use setuid() family calls at all
is for a privileged user (root) to change over to the barman
(unprivileged) user. According to POSIX.1-2013, the setuid()
call changes all three (real, effective and saved) uids if
called by a privileged user (the same holds for setgid() and
the group ids). In this construct, the additional seti[gu]id
calls are redundant no-ops and can be removed.