Fix exception when streaming connection is unavailable
Previously, when 'streaming_archiver' was on and pg_receivexlog correctly
installed, Barman would raise an AttributeError exception when streaming
connection to PostgreSQL was unavailable, as it could not detect the
server version.

This commit fixes the error by modifying the management of a failed
connection to PostgreSQL through streaming connection. It also adds
an unit test for this scenario, as well as hints for the user during
`barman check` suggesting a possible streaming connection error.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>