Add 'receive-wal' cron integration
Add integration of the new command 'receive-wal' during the execution of
the cron command.

If the option 'streaming_archiver' is on, the command 'receive-wal' is
automatically executed as subprocess by the cron command, allowing
Barman to receive WAL files from PostgreSQL using the native streaming
protocol.

TODO: Implement a way to interrupt a running streaming process

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>