Check for 'backup_label' on the master
If a PostgreSQL server is abruptly shut down when an
exclusive backup is in progress, the 'backup_label' file
will stay inside PGDATA.

In case of a failed exclusive backup, if the Postgres server
is down, "barman check" verifies that a backup_label is present.
If so, prints "A 'backup_label' is present".

Fixes #69

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>