Add a warning message to show-backup checking timelines
If after a backup there is a timeline switch, the "Last WAL" field from
the show-backup subcommand will show the last available WAL on the
timeline on which the backup was taken and not the last archived one.
This behaviour, while it's correct, can confuse the user.

Now barman handles such situation and emits a warning if there is more
than one timeline reachable from the displayed backup.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>