Check tablespace-mapping option in pg_basebackup
With `postgres` backup method, tablespaces are supported only
for PostgreSQL servers 9.3 or higher and pg_basebackup 9.4 or higher.

The reason is that pg_basebackup supports the `tablespace-mapping`
option only from 9.4, and this option is needed in case there are
tablespaces in the server. Given that pg_basebackup 9.4 is not
compatible with PostgreSQL 9.1 and 9.2, these versions are not
supported in case of tablespace presence.

This patch adds a check for this scenario, providing users with
warnings and preventing them from starting a backup in case
the aforementioned requirements are not met.

Documentation has been modified accordingly.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>