Add support for replication slots in `receive-wal` command
If `slot_name` option is set and `streaming_archiver` is enabled, the
`receive-wal` command will use the given replication slot to ensure
that no WAL file is lost before its archival.

The `check` command will monitor the status of the configured slot
on the PostgreSQL side, accordingly with the value of the
`streaming_archiver` option: if `streaming_archiver` is on, it will make
sure that the slot exists and it is used, otherwise it will warn the
user in case a stale slot is found.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>