Improve error messages for receive-wal subcommand
Error messages for the receive-wal subcommands were incoherent and this
patch improves them, using a common form.

In addition to that, the availability of the 'slot_name' configuration
parameter was checked before the server version, making that check
useless if the PostgreSQL server version has no support for replication
slots.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>