Fix error messages during create and drop of replication slots
Modified error message emitted in case of an already existing
replication slot during the execution of receive-wal --create-slot.  the
message have been changed to error from info, replicating the behaviour
of pg_receivexlog. Now, in case on error, the barman process exit code
will be 1.

The error message emitted during the execution of the receive-wal
--drop-slot in case of a not existing replication slot have been changed
to error as well.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>