Enable streaming-only archiver with replication slots
Remove the previous restriction that required standard WAL archiver
(`archiver=on`) to be always present. Allow a server to operate
with just WAL streaming if a replication slot is set (PostgreSQL
9.4 or higher is required). Replication slots make sure that every
WAL file is streamed to the Barman server.

Basic checks have been implemented at configuration level, requiring
`streaming_archiver=on` and a `slot_name` if `archiver=off`.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>