Improve replication slot error messages
Barman didn't catch connection related exceptions in the
`receive-wal` subcommand while creating and deleting a replication
slot.

This patch fixes that and reports to the user the connection error,
avoiding to print a stack trace.

This commit also adds a specific error message in case of drop of a
replication slot that is currently in use.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>