Introduce the CopyController module for backup
This patch adds a new module called copy_controller.py.

The new module is responsible for containing the logic of
a physical copy of for a backup, starting with basic
rsync support.

This patch is a refactor of the existing code, with
no additional functionality yet. Only backup executor
is using this class for now.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>