First stub of documentation for 2.0
This is a first attempt to change the documentation in Barman,
which creates two artifacts: a manual and a quickstart guide.

The work is currently incomplete, but it is ok for a first
alpha release, given that all required information from a
technical point of view is in man pages.

The reason for this is that now Barman does not have a sequential
installation procedure, but different options for backup and
WAL archiving that produce several possibile combinations of
setups.

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>