Set default value of 'archiver' to 'off'
This is a milestone in Barman's history as it
removes the need for standard WAL archiving through
the `archive_command` configuration option in
PostgreSQL (even though it is still recommended
to use it in conjunction with WAL streaming, for
more robust architectures).

**IMPORTANT**: When upgrading from 1.X.X versions
of Barman to 2.X versions, users must explicitly add
`archiver = on` (either globally or at server level).

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>