Switch xlog after a concurrent backup
For a backup to be usable, all the XLOG files written between the backup
start and the backup end must be archived.

This patch makes PostgreSQL switch the current XLOG file after a
concurrent backup either using rsync or streaming backup mode, in this
way the WAL file containing the backup end position will be archived as
soon as possible.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>