Force special tox environments to use a specific Python version
Use Python 2.7 in the 'flake8' environment, because it doesn't work with
Python 2.6.

Use Python 2.6 in the 'minimal' envorinment, because it doesn't work
with Python 3.x due to the presence of Python-2-only modules.

Enable the minimal target in Travis-CI again, as it was failing due to
the default Python version used there (Python 3.5).

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>