Add --archive and --archive-timeout options to switch-xlog command
Improve the switch-xlog command adding the --archive and
--archive-timeout options.

The --archive option will start an archiver process for the selected
server, and will wait for 30 seconds for a new xlog to be archived.

The --archive-timeout TIMEOUT will allow the user to increase or
decrease the number of seconds the archiver process will wait
for the archival of a new xlog file, before exiting.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>