Reduce time ServerXLOGDBLock is held while archiving
The ServerXLOGDBLock was held from the start to the end of the
archiving process, but that was not how it was supposed to be.

The implicit contract of a waiting lock such as the ServerXLOGDBLock
requires that the lock is kept for the shortest possible time.
However, that is not the case when archiving a large number of WAL
files or when executing the pre/post archive hooks.

This commit changes the acquisition strategy to acquire and
release the lock on a per-WAL file basis.

Fixes #99

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>