Use a process pool to execute backup and recovery
Add initial multiprocessing infrastructure to support parallel backup
and recovery using rsync.

Added the 'parallel_jobs' global/server configuration option,
that specifies the number of parallel processes used for both
backup and recovery operations. Default value is 1 (for
retro-compatibility).

Added also the `--jobs` and `-j` run-time option for the `backup`
and `recover` commands to override the server behaviour.

At the moment, when the value is set to 2 or more, copy operations
are parallelized, but the logic to split them in smaller chunks is not
yet implemented.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>