Ensure incoming directory is empty when `archiver=off`
This commit adds a check that warns the user when there are files in
the incoming directory but the `archiver` option is disabled.

This is something that could happen when the archiver is disabled in
favor of the `streaming_archiver`, but PostgreSQL is still configured to
ship WAL files to the incoming directory.

This check is a non blocking check with the purpose of warning the
user about the disk filling up with unarchived WALs.

Closes: #80

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>