Support dynamic WAL size
Barman now asks PostgreSQL for the current WAL file size at every start of
backup. This information is stored in the backup info and then used while
showing statistics.

Previously the size of WAL files was hardcoded into Barman, giving wrong
statistics when PostgreSQL was compiled with a different value.

IMPORTANT: Release notes must mention that the format of the backup info
file has changed. A backup made with Barman 2.2 cannot be read by a
previous version of Barman (in the rare event that a user downgrades
version).

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.com>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>