Make cuncurrent backup killable
Barman was able to recover from a KeyboardInterrupt while
executing a backup, so you were not able to interrupt a backup anymore.

This patch disable the handling of the SIGINT signal from the worker processes.
The signal is now handled by the parent process, which is responsible
for the termination of the children processes.
The backup will be then stopped in the usual way.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.com>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.com>