Check the number of files in the archiving queue
Add a new configuration parameter, named `max_incoming_wals_queue`, that
enables a new check for the number of files in the incoming directories.

This check will fail if the number of WALs to be archived
goes over the allowed threshold.

This check does not block normal operations.

The purpose of this check is to improve robustness of the Barman
system when integrated in an alerting/monitoring infrastructure,
by warning users in case Barman is not archiving fast enough.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.com>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>