Ignore `.tmp` files in `incoming` and `streaming` directories
The pg_receivewal utility occasionally writes to a `.tmp` file and
then renames it to the final name to ensure an atomic creation.
Barman must ignore this temporary file.
Other programs could do the same, so Barman will ignore any file name
with a `.tmp` suffix in the incoming and streaming directories.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>