Stop process manager module from ovewriting lock files content
Previously a bug in the process manager code was writing its own PID
in all the existing lock files which didn't have an owner.

Due to a bug, the content of all the lock files without owner
was overwritten with the PID of every barman command interacting with
a server object.

Because of this barman cron was wrongly identifying other commands
as a running `receive-wal` process and then terminating them.

The code has now changed to avoid modifying the content of the lock
file when accessed by the process manager.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>