Improve usability of point-in-time recovery
Add a check to stop the recovery immediately if the recovery time is
before the backup end time.

If the timezone is not specified, the target recovery time is
interpreted according to the Barman server timezone. This is a
behavior change, because previously the specified time was simply
passed to PostgreSQL and thus interpreted according the PostgreSQL
settings.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.com>