i18n for barman
This commit adds the internationalization support to barman. To create a new language file follow the steps below.

For example to create the translation for the Italian language:

$ python setup.py build
$ cp po/barman.pot po/it.po

Register the new language by modifying in the file barman/i18n.py

LANGUAGES = [
    # Add languages here
    ('es', 'Spanish'),
    ('es', 'Italian'),
]

In addition, the construction process was extended, for the compilation and creation of * .mo files