Add `check-backup` command
Make sure that all the required WAL files to check the consistency
of a physical backup (that is, from the beginning to the end of
the full backup, as outlined in the backup label) are correctly archived.
This command is automatically invoked by the `cron` command,
and at the end of every backup operation.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>