Geographic redundancy implementation
Introduce three new commands (`sync-info`, `sync-backup` and `sync-wals`)
and one global/server configuration option called `primary_ssh_command`.

When the latter is specified globally, the whole instance of Barman is
an async copy of the server reached via SSH by `primary_ssh_command`.
If specified on a single server definition, that server in Barman is
an async copy of the same server that is defined on another Barman
installation.

Geo-redundancy is asynchronous and based on SSH connections between
Barman servers. Cascading backup is supported.

Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>