Fix "Backup completed" elapsed time
The message issued when a backup completes does not include all of the
elapsed time, but rather only a non-intuitive subset of it.

Notably it does not include the time taken to fsync all of the files,
which can be substantial. Make the message include all time up to when
the message is issued.

Pull request GH-216

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndQuadrant.it>