Fix exclude_and_protect_filter
Filters which do not start with '/' are treated as wildcards.
For the exclude_and_protect_filter, this is both slow and
incorrect. For example, if there were a spurious file named
"pg_xact/0000" somewhere other than the top-level dest data
directory, the unachored protect filter means it would not
get deleted.  With an anchored protect filter, it will get
deleted, as it should.

So fix this by prefixing the filter paths with a '/'.

The bug fix is minor (it is not clear why such spurious files
would exist in the first place) but the speed-up should be
appreciated by anyone with a large number of files in their
database.

Closes: GH-217

Signed-off-by: Jeff Janes <jeff.janes@gmail.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>