Skip WAL archive failure if we have a terminated backup
In low workload situations with concurrent backup, at the end of the
backup the WAL archive might be temporarily empty and therefore throw
a failure with `check` or commands that rely on it such as `backup`.
Ignore the error if the last backup is `WAITING_FOR_WALS` as WAL
archiving must have been verified before running that backup.

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>