Truncate the .partial WAL file when size is wrong
During the `receive-wal` startup, if we detect that the partial file size
is wrong, we truncate it and then start `pg_receivewal`, making Barman
more resilient.

This is coherent with the behaviour of `pg_receivewal`, that will allocate
the exact size needed for a WAL segment before starting to write it.
This also enables `receive-wal` to automatically resume after the backup
server ran out of disk space.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>