Ignore vanished and truncated files uploading cloud backups
This is to avoid backup errors if PostgreSQL rewrite or delete a file in
its data directory while the backup is running.

If a file change its size during the backup, we pad the file contents
with zeroes, and the WAL reply will take care of the original contents
during recovery.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>