Add support for .partial WAL files to `get-wal`
Remove the long-standing limitation of `restore_command` that
were not capable of automatically handling `.partial` WAL files
shipped to Barman via streaming replication.
When invoked with `--partial/-P`, `get-wal` command will look
also for the `.partial` content of the requested WAL file.
This is extremely useful in case of disaster recovery situations
with RPO=0.

The patch also covers `barman-wal-restore`.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>