Support arbitrary archive size in barman-cloud-backup
S3 has some limitations on the maximum size of a single file.
For this reason, when uploading large tar files, we need to split
them in multiple files.

A new option to `barman-cloud-backup` called `--max-archive-size`,
by default set to 100GB, to split a tar file in multiple parts
when that size is reached.

Signed-off-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Signed-off-by: Giulio Calacoci <giulio.calacoci@2ndquadrant.it>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@2ndQuadrant.it>