Fetch wal_keep_size, not wal_keep_segments, from Postgres 13
The `wal_keep_segments` parameter introduced in v9.0 was replaced with
`wal_keep_size` in v13. Running Barman against Postgres 13 resulted in
errors like the following being logged:

    barman ERROR:  unrecognized configuration parameter "wal_keep_segments"
    barman STATEMENT:  SHOW "wal_keep_segments"

Here we change fetch_remote_status() to ask for wal_keep_size if the
server version is >= 13. We didn't use this value anywhere, so we don't
need any further changes to adapt to the different return value.

Signed-off-by: Abhijit Menon-Sen <ams@2ndQuadrant.com>