Stop backups failing due to idle transactions
Uses the psycopg2 connection context manager so that transactions
are committed when barman uses the connection to interact with
postgres for reasons other than starting or stopping the backup.

This prevents the following situation from arising:

1. Barman starts a backup by SELECTing from pg_start_backup and
   immediately rolls back the transaction.
2. Barman starts a new transaction to retrieve the server
   version.
2. The transaction is left open putting the session into an
   idle-in-transaction state.
3. The backup lasts longer than the value of
   idle_in_transaction_session_timeout and fails because the
   session is terminated.

Because this patch ensures such transactions are committed,
via the psycopg2 connection context manager, such transactions
are no longer held open during the backup.

Note that in cases where we do not call `self.connect()` directly,
a new context manager is added around the `_cursor` property in
barman.postgres.PostgreSQL which uses with-statement to enter the
connection context manager and then yields the cursor context
manager. When the cursor context manager exits, the connection
context manager will also exit. It is not sufficient to just use
the cursor context manager as this does not commit transactions.

See https://www.psycopg.org/docs/usage.html#with-statement for
more details.

The context manager is deliberately avoided for the functions
which call either `pg_{start,stop}_backup` or
`pgespresso_{start,stop}_backup` as these manage transactions
themselves by explicitly calling `rollback`.

Closes #333