Refactor barman-cloud (1/6): Generalise CloudInterface
This is commit 1 of 6 where barman-cloud is refactored so that
support can be added for additional cloud providers.

Generalises CloudInterface into an abstract base class which
provides the logic for asynchronously uploading large objects
concurrently as well as some boilerplate for interacting with
cloud storage.

The methods which interact directly with the cloud privider are
replaced with @abstractmethod stubs which must be implemented by
each provider-specific subclass.

The only current provider-specific subclass is S3CloudInterface
which now contains all the S3-specific logic previously found in
CloudInterface.

For the most part this commit simply moves chunks of code around
however the following changes were necessary to get a clean
separation between CloudInterface and the S3 implementation
details:

1. A new abstractmethod, _create_bucket, provides a home for the
   provider-specific bucket creation logic. This allows us to
   keep the "check bucket exists, create if it doesn't exist,
   set the bucket_exists flag" boilerplate in setup_bucket.
2. The extract_tar method now uses remote_open instead of
   calling S3 directly so that it can remain in CloudInterface
   and be re-used. This means we're now passing StreamingBodyIO
   (the wrapper around the boto3 StreamingBody) into tarfile
   however I don't see any downsides to doing so.
3. A new abstractmethod, reinit_session, provides a generic means
   for worker_process_main to reinitialise any session related
   resources it gets from its parent process.

We also make the following CloudInterfaces methods protected
(previously they were public but nothing is calling them
outside of CloudInterface):

* abort
* worker_process_main
* worker_process_execute_job
* check_bucket_existence
* upload_part
* complete_multipart_upload
* abort_multipart_upload