Refactor barman-cloud (2/6): Reorder CloudInterface methods
This is commit 2 of 6 where barman-cloud is refactored so that
support can be added for additional cloud providers.

Reorders methods in CloudInterface with the intention of making it
easier to read for maintainers adding support for additional cloud
providers.

The new order is as follows:

1. Methods relating to the asynchronous concurrent upload of object
   parts are first.
2. Boilerplate methods for interacting with buckets and exposing
   objects as a tarfile follow.
3. Finally, the abstract methods which must be implemented by
   sub-classes.

The implementation of the abstract methods in S3CloudInterface
follows the same order as in CloudInterface.