Refactor barman-cloud (4/6): Make cloud functions generic
This is commit 4 of 6 where barman-cloud is refactored so that
support can be added for additional cloud providers.

Replaces S3* functions with cloud-neutral Cloud* alternatives.
These are largely the same aside from the names and some
documentation because the cloud-specific details were moved to
S3CloudInterface in the previous two commits.