Refactor barman-cloud (5/6): Move S3CloudInterface to cloud_providers module
This is commit 5 of 6 where barman-cloud is refactored so that
support can be added for additional cloud providers.

Moves S3CloudInterface and associated boto3 code out of
barman.cloud and into a new cloud_providers module. This is so
that we can conditionally import the dependencies for a
specified cloud provider, rather than importing all the
dependencies for all supported cloud providers.