Refactor barman-cloud (6/6): Allow cloud provider to be specified
This is commit 6 of 6 where barman-cloud is refactored so that
support can be added for additional cloud providers.

Adds a new argument, --cloud-provider, which determines which
subclass of CloudInterface is created. Creation itself is now
handled in a factory method which imports the specified
CloudInterface on-demand and instantiates it.

The argument defaults to aws-s3 which is currently the only
supported cloud provider.