Add Azure Blob Storage support to barman-cloud
Adds a new cloud provider, azure_blob_storage, which allows
barman-cloud to be used with the Azure blob storage service.

We aim for parity with the aws_s3 cloud provider, however there are
a number of key differences between S3 and Azure blob storage which
need to be accounted for:

1. The Azure equivalent to the S3 multipart upload API works
differently. In S3 we create a multipart upload, upload individual
parts which each reference the handle for that multipart upload,
then finalize the upload by passing a list of all the part IDs
along with the multipart upload handle. In Azure the equivalent
operation is to upload individual blocks with assigned block ID
values. These only need to be associated with the blob key. The
upload is finalized by committing a list of all uploaded block IDs.
This leads to some oddities in the implementations of the
_create_multipart_upload, _upload_part, _complete_multipart_upload
and _abort_multipart_upload  abstract methods. Specifically,
_create_multipart_upload is a no-op, and the other three methods
all accept a multipart upload handle which they ignore entirely.

2. The equivalent of the boto3 StreamingBody is
azure.storage.blob.StorageStreamDownloader. Like StreamingBody this
allows us to download large objects as chunks however unlike
StreamingBody it does not act like a Python file-like object. We
therefore have to fetch chunks ourselves in our own implementation
of read() until we have read the requested number of bytes.

3. Because there is no concept of a multipart upload handle, there
is no straightforward way to clean up uploaded blocks when aborting
an upload. It is also not possible to delete individual blocks. The
only way blocks can be deleted is if they are not committed after
7 days (in which case they are automatically deleter), or if the
blob is committed and the blck is not included in the block list.
We can therefore do some basic clean up in the
_abort_multipart_upload implementation by committing an empty
block list (to discard all the staged blocks) and then deleting the
empty blob.

4. Azure blob service URIs are ambiguous because of the concept
of a root container for a storage account. This means it is not
possible to tell when parsing the URI whether the first element of
the path is referencing the name of a container or the name of a
blob in the root container. We therefore choose not to support
implicit root containers and always assume the first item in the
URI path is the container name.