Allow barman-cloud-backup to run as a hook script
Adds a new CloudBackupUploaderBarman class which uploads a
pre-existing backup taken by barman. The environment is checked
for hook script variables and if they are present the new
CloudBackupUploaderBarman class is used instead of the original
class which uploads a live backup from a Postgres server (now
renamed to CloudBackupUploaderPostgres).

This allows barman-cloud-backup to be run as a post-backup
hook script on the barman server, for cases where users wish to
take a backup with barman and then upload it to cloud storage
after the fact.

barman-cloud-backup requires that the backup has status DONE and
will error with an ABORT_STOP (exit code 63) if the backup has
any other status. This implicitly requires that backups which
trigger the barman-cloud-backup hook script are run with the
`-w / --wait` flag so that the status is not WAITING_FOR_WALS
when the hook script runs.