Allow barman-cloud-wal-archive to run as a post-archive hook script
Checks the environment in barman-cloud-wal-archive for hook script
variables and if they are present reads the wal path from
BARMAN_FILE instead of the command line.

This allows barman-cloud-wal-archive to be run as a post-archive
hook script on the barman server, for cases where users wish to
archive WALs to the barman server and from there upload them to
cloud storage.

It is necessary to be careful with the configuration of compression
options between barman and barman-cloud as compression can be
applied in both places. If this happens then the WALs uploaded to
cloud storage will be compressed twice which means
barman-cloud-wal-restore will not be able to restore them directly
from cloud storage.